feat：拆分子心流的思考模块
